Move deploy tooling back out of script
